Split BaseTransformerTest
